+++ packages/keyboard-shortcuts/src/components/shortcut-provider.js
@@ -1,0 +1,31 @@
+/**
+ * WordPress dependencies
+ */
+import { useRef } from '@wordpress/element';
+
+/**
+ * Internal dependencies
+ */
+import { context } from '../context';
+
+const { Provider } = context;
+
+export function ShortcutProvider( props ) {
+	const keyboardShortcuts = useRef( new Set() );
+
+	function onKeyDown( event ) {
+		if ( props.onKeyDown ) props.onKeyDown( event );
+
+		for ( const keyboardShortcut of keyboardShortcuts.current ) {
+			keyboardShortcut( event );
+		}
+	}
+
+	/* eslint-disable jsx-a11y/no-static-element-interactions */
+	return (
+		<Provider value={ keyboardShortcuts }>
+			<div { ...props } onKeyDown={ onKeyDown } />
+		</Provider>
+	);
+	/* eslint-enable jsx-a11y/no-static-element-interactions */
+}
