+++ packages/keyboard-shortcuts/src/context.js
@@ -1,0 +1,6 @@
+/**
+ * WordPress dependencies
+ */
+import { createContext } from '@wordpress/element';
+
+export const context = createContext();
