+++ packages/keyboard-shortcuts/src/index.js
@@ -1,3 +1,4 @@
 export { store } from './store';
-export { default as useShortcut, ShortcutProvider } from './hooks/use-shortcut';
+export { default as useShortcut } from './hooks/use-shortcut';
+export { ShortcutProvider } from './components/shortcut-provider';
 export { default as __unstableUseShortcutEventMatch } from './hooks/use-shortcut-event-match';
